Argh, I tested out the PID control with limited success
I've discovered and fixed a bunch of errors having to do with the PID
and motor driver codes. I still have a bunch of bugs listed. I need to
contain a lot of what I'm doing into a class to simplify the program. I
also need to learn to write tester programs.