I've added a new class: PIDController and I've started tuning the PID
I'm not completely satisfied with the PIDController class. However, I've
got it working and I'd like to try to get my PID working. I think it's
time to work on an auto tuner for the PID. I could spend forever trying
to hand tweak the parameters. I also need to learn more about what
exactly the parameters are in my PID.