I modified the PID class to enable turning off some features and fixed the way it used outputs
The class will now turn of windup and sensor filtering if their
respective inputs are set to zero. Before, the filtering equation would
explode if fiterCoef was zero. Now it will just turn off filtering. Same
thing with the windup equation. Additionally, the PID equations will now
use the most recent output sent to the system for their windup
calculations instead of the most recently calculated. This is a fairly
subtle difference.